requirements.txt
progressbar
elasticsearch
pandas
pandarallel
tenacity
tqdm
requests
GitPython
javalang
tokenizers
transformers
datasets